backend/pyproject.toml
[project]
name = "agriflow-backend"
version = "0.1.0"
description = "FastAPI backend for AgriFlow crop recommendations"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14,<2.0",
  "fastapi>=0.115,<1.0",
  "httpx>=0.28,<1.0",
  "joblib>=1.4,<2.0",
  "pandas>=2.2,<3.0",
  "scikit-learn>=1.4,<2.0; python_version < '3.13'",
  "pydantic-settings>=2.8,<3.0",
  "psycopg[binary]>=3.2,<4.0",
  "sqlalchemy>=2.0,<3.0",
  "structlog>=25.1,<26.0",
  "uvicorn[standard]>=0.34,<1.0",
]

[project.optional-dependencies]
dev = [
  "anyio>=4.8,<5.0",
  "pytest>=8.3,<9.0",
  "pytest-cov>=6.0,<7.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"
